Add files via upload
It is a simple  CORE JAVA project .  In this project I perform crude operations. it is easy to understand and use.
In this proeject i provide manu for user and according that manu user can easily hand this project .
In this project user can update the car details, delete and insert the details.